refactor: registration works
