+++ src/bot/registration/__init__.py
@@ -1,0 +1,5 @@
+from .registration_router import include_registration_routers
+
+__all__ = [
+    "include_registration_routers",
+]
