+++ src/bot/registration/resources/reply_buttons.py
@@ -1,7 +1,7 @@
 from aiogram.types import KeyboardButton, ReplyKeyboardMarkup
 from aiogram.utils.keyboard import ReplyKeyboardBuilder
 
-from src.bot.common.command_filter import TelegramCommand
+from src.bot.common import TelegramCommand
 
 __all__ = [
     "start_button",
