+++ src/bot/root_router.py
@@ -1,6 +1,7 @@
 from .common import RootRouter
 from .help import include_help_command_router
 from .poll_attendance import include_poll_attendance_routers
+from .registration import include_registration_routers
 from .show_group_attendance import include_show_group_attendance_routers
 
 __all__ = [
@@ -10,6 +11,7 @@ __all__ = [
 def build_root_router() -> RootRouter:
     root_router = RootRouter(throttling=True)
 
+    include_registration_routers(root_router)
     include_help_command_router(root_router)
     include_poll_attendance_routers(root_router)
     include_show_group_attendance_routers(root_router)
